+++ src/week1/domain/CustomProduct.java
@@ -1,4 +1,4 @@
-package domain;
+package week1.domain;
 
 import java.util.ArrayList;
 import java.util.List;
