+++ src/week1/domain/Salad.java
@@ -1,4 +1,4 @@
-package domain;
+package week1.domain;
 
 public class Salad extends CustomProduct{
     int price;
