+++ src/week1/domain/Vegetable.java
@@ -1,12 +1,11 @@
-package domain;
+package week1.domain;
 
-public class Product {
+public class Vegetable {
     String name;
 
-    public Product (String name){
+    public Vegetable(String name){
         this.name = name;
     }
-
     public String getName() {
         return name;
     }
